Moved Kotlin files out of kotlin subpackage
