+++ app/src/main/java/cm/aptoide/pt/autoupdate/AutoUpdateModel.kt
@@ -1,9 +1,8 @@
-package cm.aptoide.pt.autoupdate.kotlin
+package cm.aptoide.pt.autoupdate
 
 
 data class AutoUpdateModel(val versionCode: Int, val uri: String, val md5: String,
-                           val minSdk: String, val packageName: String, val shouldUpdate: Boolean = false,
-                           var error: Error? = null, var loading: Boolean = false) {
+                           val minSdk: String, val packageName: String, val shouldUpdate: Boolean = false, var error: Error? = null, var loading: Boolean = false) {
 
     constructor(error: Error?) : this(-1, "", "", "", "", error = error)
