+++ app/src/main/java/cm/aptoide/pt/autoupdate/AutoUpdateRepository.kt
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.autoupdate.kotlin
+package cm.aptoide.pt.autoupdate
 
 import rx.Single
