+++ app/src/main/java/cm/aptoide/pt/view/ActivityModule.java
@@ -27,10 +27,10 @@ import cm.aptoide.pt.ads.AdsRepository;
 import cm.aptoide.pt.app.AppNavigator;
 import cm.aptoide.pt.app.view.AppViewNavigator;
 import cm.aptoide.pt.app.view.donations.DonationsAnalytics;
-import cm.aptoide.pt.autoupdate.kotlin.AutoUpdateManager;
-import cm.aptoide.pt.autoupdate.kotlin.AutoUpdateRepository;
-import cm.aptoide.pt.autoupdate.kotlin.AutoUpdateService;
-import cm.aptoide.pt.autoupdate.kotlin.Service;
+import cm.aptoide.pt.autoupdate.AutoUpdateManager;
+import cm.aptoide.pt.autoupdate.AutoUpdateRepository;
+import cm.aptoide.pt.autoupdate.AutoUpdateService;
+import cm.aptoide.pt.autoupdate.Service;
 import cm.aptoide.pt.crashreports.CrashReport;
 import cm.aptoide.pt.database.accessors.StoreAccessor;
 import cm.aptoide.pt.dataprovider.interfaces.TokenInvalidator;
